requirements.txt
asgiref==3.7.2
bleach==6.0.0
blessed==1.20.0
boto3==1.28.35
botocore==1.31.35
bpython==0.24
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
coverage==7.2.7
crispy-bootstrap5==0.7
cryptography==41.0.3
cssbeautifier==1.14.9
curtsies==0.4.2
cwcwidth==0.1.8
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.20
django-allauth==0.54.0
django-crispy-forms==2.0
django-extensions==3.2.3
django-storages==1.13.2
django-summernote==0.8.20.0
djlint==1.32.1
EditorConfig==0.12.3
factory-boy==3.3.0
Faker==19.3.1
greenlet==2.0.2
gunicorn==21.2.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
jmespath==1.0.1
jsbeautifier==1.14.9
json5==0.9.14
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
psycopg2-binary==2.9.7
pycparser==2.21
Pygments==2.15.1
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
pyxdg==0.28
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
stripe==5.5.0
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==1.26.16
wcwidth==0.2.6
webencodings==0.5.1
